# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Luminae multiplayer tabletop game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/architect-record:
    get:
      operationId: getArchitectRecord
      tags: [architect-record]
      summary: Fetch the account's redacted Architect Record
      security:
        - accountToken: []
      responses:
        "200":
          description: Redacted Architect Record state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectRecordState"
  /campaigns/architect-record/onboarding-claim:
    post:
      operationId: claimArchitectRecordOnboarding
      tags: [architect-record]
      summary: Idempotently claim a completed first-contact tutorial
      security:
        - accountToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingClaimBody"
      responses:
        "200":
          description: Updated Architect Record state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectRecordState"
  /campaigns/architect-record/presentations/{presentationId}/acknowledge:
    post:
      operationId: acknowledgeArchitectRecordPresentation
      tags: [architect-record]
      summary: Idempotently acknowledge a saved Architect Record presentation
      security:
        - accountToken: []
      parameters:
        - name: presentationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Presentation acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /campaigns/architect-record/qualifying-match:
    post:
      operationId: createOrResumeQualifyingMatch
      tags: [architect-record]
      summary: Create or resume a standard 15-Eminence match against three Hard AIs
      security:
        - accountToken: []
      responses:
        "200":
          description: Existing qualifying match resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualifyingMatchSession"
        "201":
          description: New qualifying match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualifyingMatchSession"
  /events/first-party:
    post:
      operationId: recordFirstPartyEvent
      tags: [telemetry]
      summary: Accept an allowlisted event with fixed, non-personal dimensions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirstPartyEventBody"
      responses:
        "202":
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                required: [accepted]
...
components:
  securitySchemes:
    accountToken:
      type: http
      scheme: bearer
  schemas:
...
    FirstContactStance:
      type: string
      enum: [curious, guarded, resolute]
    CampaignNodeStatus:
      type: string
      enum: [locked, available, active, completed, future]
    CampaignPresentation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [clearance_signal, clearance_recap]
        ordinal:
          type: integer
        title:
          type: string
        lines:
          type: array
          items:
            type: string
        acknowledgedAt:
          type: [string, "null"]
          format: date-time
      required: [id, kind, ordinal, title, lines, acknowledgedAt]
    CampaignNodeSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        title:
          type: string
        status:
          $ref: "#/components/schemas/CampaignNodeStatus"
        progress:
          type: integer
        requiredProgress:
          type: integer
      required: [id, title, status, progress, requiredProgress]
    ArchitectRecordState:
      type: object
      additionalProperties: false
      properties:
        campaignId:
          type: string
          enum: [architect_record]
        tutorialCompleted:
          type: boolean
        firstContactStance:
          anyOf:
            - $ref: "#/components/schemas/FirstContactStance"
            - type: "null"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CampaignNodeSummary"
        presentations:
          type: array
          items:
            $ref: "#/components/schemas/CampaignPresentation"
        pendingPresentations:
          type: array
          items:
            $ref: "#/components/schemas/CampaignPresentation"
        vaultShortcutVisible:
          type: boolean
      required: [campaignId, tutorialCompleted, firstContactStance, nodes, presentations, pendingPresentations, vaultShortcutVisible]
    OnboardingClaimBody:
      type: object
      additionalProperties: false
      properties:
        claimId:
          type: string
          format: uuid
        stance:
          anyOf:
            - $ref: "#/components/schemas/FirstContactStance"
            - type: "null"
      required: [claimId, stance]
    QualifyingMatchSession:
      type: object
      additionalProperties: false
      properties:
        roomId:
          type: string
          format: uuid
        inviteCode:
          type: string
        playerId:
          type: string
          format: uuid
        sessionToken:
          type: string
        playerName:
          type: string
        resumed:
          type: boolean
      required: [roomId, inviteCode, playerId, sessionToken, playerName, resumed]
    FirstPartyEventBody:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        anonymousSessionId:
          type: [string, "null"]
          format: uuid
        eventName:
          type: string
          enum: [tutorial_started, tutorial_resumed, tutorial_restarted, tutorial_left, tutorial_chapter_started, tutorial_chapter_completed, tutorial_invalid_action, tutorial_completed, account_prompt_outcome, guided_practice_started, qualification_milestone, interlude_acknowledged, vault_completed, turn_order_balance_viewed, match_balance_result]
        chapterId:
          type: [string, "null"]
          enum: [arrival, board, actions, ascension, null]
        beatId:
          type: [string, "null"]
          maxLength: 64
        actionId:
          type: [string, "null"]
          enum: [harness, forge, encrypt, forge_reserved, forge_final, help, register, sign_in, dismiss, match_2p_v15, match_2p_v20, match_2p_v25, match_3p_v15, match_3p_v20, match_3p_v25, match_4p_v15, match_4p_v20, match_4p_v25, null]
        outcome:
          type: [string, "null"]
          enum: [allowed, blocked, success, failure, accepted, dismissed, resume, start_over, null]
        ordinal:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
        durationMs:
          type: [integer, "null"]
          minimum: 0
          maximum: 86400000
        occurredAt:
          type: string
          format: date-time
      required: [id, eventName]
...